2PC-Microservice/services/ui/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
grpcio==1.66.1
protobuf==5.27.3
PyJWT==2.9.0
python-multipart==0.0.9
python-multipart
